test cases

+++ prepare/Learning03/Fraction.cs
@@ -50,6 +50,6 @@ class Fraction
 
     public double GetDecimalValue()
     {
-        return _top/_bottom;
+        return (double)_top/(double)_bottom;
     }
 }
